+++ changelog.md
@@ -37,6 +37,7 @@ Codion Change Log
 - EntityComponents.create() added, related refactoring.
 - EntityEditPanel no longer implements EntityComponents.
 - DefaultEditorComponents.DefaultEditorComponent.set() bug fixed, disables more carefully in case of derived attributes, previously interfered with initial action based text field enabled state.
+- EntityEditPanel now returns the focus to the last focused component after update, not last focused input component.
 
 ## 0.18.69
 ### is.codion.common.utilities